Move env types to global.d.ts
