+++ tsconfig.json
@@ -17,8 +17,7 @@
     "./plugins/**/*",
     "./gatsby-node.ts",
     "./gatsby-config.ts",
-    "./overrides.d.ts",
-    "./environment.d.ts"
+    "./overrides.d.ts"
   ],
   "exclude": [
     "node_modules",
